requirements.txt
alabaster==0.7.12
Babel==2.6.0
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
docutils==0.14
idna==2.7
imagesize==1.1.0
IMDbPY==6.6
Jinja2==2.10
lxml==4.2.5
MarkupSafe==1.1.0
nltk==3.3
packaging==18.0
Pygments==2.2.0
pyparsing==2.3.0
-e git+https://github.com/agonzalezro/python-opensubtitles@66a5f893912e70fe9884eb1115be0ffe6a4d31fa#egg=python_opensubtitles
pytz==2018.7
requests==2.20.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.2
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
SQLAlchemy==1.2.14
urllib3==1.24.1
